+++ mantis/__version__.py
@@ -1,1 +1,7 @@
 __version__ = '0.1.0'
+__name__="mantis"
+__description__="Mantis CLI and Library for NLP utils"
+__url__="https://github.com/MantisAI/mantis"
+__author__="Mantis NLP"
+__author_email__=""
+__license__=""
