+++ requirements.txt
@@ -3,5 +3,7 @@ typer
 spacy
 transformers
 sentence-transformers
+tensorflow
 tensorflow-hub
 pandas
+boto3
